flow to show wishlist items in the page recyclerView

+++ app/src/main/java/com/example/obes/MyAdapterRecyclerViewCart.java
@@ -42,7 +42,7 @@ public class MyAdapterRecyclerViewCart extends RecyclerView.Adapter<MyAdapterRec
         if (this.isShoppingCart) {
             view = LayoutInflater.from(context).inflate(R.layout.item_cart, parent, false);
         } else {
-            view = LayoutInflater.from(context).inflate(R.layout.item_cart, parent, false);
+            view = LayoutInflater.from(context).inflate(R.layout.item_wishlist, parent, false);
         }
         return new MyAdapterRecyclerViewCart.MyHolder(view);
     }
@@ -72,24 +72,26 @@ public class MyAdapterRecyclerViewCart extends RecyclerView.Adapter<MyAdapterRec
             }
         });
 
-        holder.ItemCart.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                Activity activity = (Activity) context;
-                TextView tvPriceTotal = activity.findViewById(R.id.price_total);
-                DecimalFormat df = new DecimalFormat("#.00");
-                String newPriceTotal;
+        if (this.isShoppingCart) {
+            holder.ItemCart.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+                @Override
+                public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
+                    Activity activity = (Activity) context;
+                    TextView tvPriceTotal = activity.findViewById(R.id.price_total);
+                    DecimalFormat df = new DecimalFormat("#.00");
+                    String newPriceTotal;
 
-                if (isChecked) {
-                    priceTotal = countPriceTotal(book.getPrice());
-                } else {
-                    priceTotal = countPriceTotal(book.getPrice() * -1);
+                    if (isChecked) {
+                        priceTotal = countPriceTotal(book.getPrice());
+                    } else {
+                        priceTotal = countPriceTotal(book.getPrice() * -1);
+                    }
+
+                    newPriceTotal = "R$ " + (priceTotal > 0 ? df.format(priceTotal) : "0.00");
+                    tvPriceTotal.setText(newPriceTotal);
                 }
-
-                newPriceTotal = "R$ " + (priceTotal > 0 ? df.format(priceTotal) : "0.00");
-                tvPriceTotal.setText(newPriceTotal);
-            }
-        });
+            });
+        }
 
         holder.ivDelete.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -103,15 +105,17 @@ public class MyAdapterRecyclerViewCart extends RecyclerView.Adapter<MyAdapterRec
 
                     notifyItemRemoved(itemPosition);
 
-                    priceTotal = countPriceTotal(book.getPrice() * -1);
+                    if (isShoppingCart) {
+                        priceTotal = countPriceTotal(book.getPrice() * -1);
 
-                    updatePriceTotalInUI();
+                        updatePriceTotalInUI();
 
-                    ItemCart item = itemCartDAO.getItemByIdBook(book.getId());
-                    if (item != null) {
-                        itemCartDAO.deleteItemCart(item);
-                        int idCart = CartToItemDAO.getInstance().getIdCartByIdItem(item.getId());
-                        CartToItemDAO.getInstance().deleteCartItem(idCart, item.getId());
+                        ItemCart item = itemCartDAO.getItemByIdBook(book.getId());
+                        if (item != null) {
+                            itemCartDAO.deleteItemCart(item);
+                            int idCart = CartToItemDAO.getInstance().getIdCartByIdItem(item.getId());
+                            CartToItemDAO.getInstance().deleteCartItem(idCart, item.getId());
+                        }
                     }
                 }
             }
